[feat] #26 이용예정/완료, 대기상세 뷰 연결

+++ app/src/main/java/org/sopt/tabling/data/service/WaitingDetailService.kt
@@ -6,5 +6,5 @@ import retrofit2.http.Path
 
 interface WaitingDetailService {
     @GET("/orders/{order_id}")
-    suspend fun getWaitingDetail(@Path("order_id") orderId: Int): ResponseWaitingDetailDto
+    suspend fun getWaitingDetail(@Path("order_id") orderId: Long): ResponseWaitingDetailDto
 }
